+++ src/Paradigm.Enterprise.Services.Cache/Configuration/RedisCacheConfiguration.cs
@@ -4,4 +4,5 @@ internal class RedisCacheConfiguration
 {
     public int? ExpirationInMinutes { get; set; }
     public bool Disabled { get; set; }
+    public RedisManagedIdentityConfiguration? ManagedIdentity { get; set; }
 }