+++ src/Paradigm.Enterprise.Services.Email/Models/EmailClientCreationStrategy.cs
@@ -1,0 +1,8 @@
+﻿namespace Paradigm.Enterprise.Services.Email.Models;
+
+internal enum EmailClientCreationStrategy
+{
+    Invalid = 0,
+    ConnectionString = 1,
+    ManagedIdentity = 2,
+}